+++ packages/executor/src/execution/__tests__/error-handling.test.ts
@@ -35,16 +35,20 @@ describe('Error Handling', () => {
     expect(result.errors?.[0]?.message).toBe('This is not an error instance');
   });
   if (globalThis.fetch != null) {
-    let server: Server;
+    let server: Server | undefined;
     afterEach(async () => {
-      if (!server) {
+      if (!server?.listening) {
         return;
       }
       if (!globalThis.Bun) {
         server.closeAllConnections();
       }
       await new Promise<void>((resolve, reject) => {
+        if (!server) {
+          return resolve();
+        }
         server.close(err => (err ? reject(err) : resolve()));
+        server = undefined;
       });
     });
     it('handles undici fetch JSON parsing errors', async () => {
@@ -52,6 +56,9 @@ describe('Error Handling', () => {
         res.end('{ "myData": "foo"');
       });
       await new Promise<void>(resolve => {
+        if (!server) {
+          throw new Error('Server is not initialized');
+        }
         server.listen(0, resolve);
       });
       const serverPort = (server.address() as AddressInfo).port;
@@ -94,4 +101,60 @@ describe('Error Handling', () => {
       }
     });
   }
+  it('handles aggregated errors', async () => {
+    const schema = makeExecutableSchema({
+      typeDefs: /* GraphQL */ `
+        type Query {
+          throwMe: String
+        }
+      `,
+      resolvers: {
+        Query: {
+          throwMe: () =>
+            new AggregateError(
+              [new Error('This is an error'), new Error('This is another error')],
+              'This is an aggregated error',
+            ),
+        },
+      },
+    });
+    const result = await normalizedExecutor({
+      schema,
+      document: parse(/* GraphQL */ `
+        query {
+          throwMe
+        }
+      `),
+    });
+    if (isAsyncIterable(result)) {
+      throw new Error('Expected a result, but got an async iterable');
+    }
+    expect(JSON.parse(JSON.stringify(result))).toEqual({
+      data: {
+        throwMe: null,
+      },
+      errors: [
+        {
+          locations: [
+            {
+              column: 11,
+              line: 3,
+            },
+          ],
+          message: 'This is an error',
+          path: ['throwMe'],
+        },
+        {
+          locations: [
+            {
+              column: 11,
+              line: 3,
+            },
+          ],
+          message: 'This is another error',
+          path: ['throwMe'],
+        },
+      ],
+    });
+  });
 });