+++ packages/executors/urql-exchange/tests/urql-exchange.spec.ts
@@ -4,10 +4,12 @@ import { pipe, toObservable } from 'wonka';
 import { buildHTTPExecutor } from '@graphql-tools/executor-http';
 import { ExecutionResult } from '@graphql-tools/utils';
 import { createClient } from '@urql/core';
+import { AsyncDisposableStack } from '@whatwg-node/disposablestack';
 import { testIf } from '../../../testing/utils.js';
 import { executorExchange } from '../src/index.js';
 
 describe('URQL Yoga Exchange', () => {
+  const asyncDisposableStack = new AsyncDisposableStack();
   const yoga = createYoga({
     logging: false,
     maskedErrors: false,
@@ -26,10 +28,10 @@ describe('URQL Yoga Exchange', () => {
       `,
       resolvers: {
         Query: {
-          hello: () => 'Hello Urql Client!',
+          hello: async () => 'Hello Urql Client!',
         },
         Mutation: {
-          readFile: (_, args: { file: File }) => args.file.text(),
+          readFile: async (_, args: { file: File }) => args.file.text(),
         },
         Subscription: {
           alphabet: {
@@ -48,20 +50,21 @@ describe('URQL Yoga Exchange', () => {
       },
     }),
   });
+  asyncDisposableStack.use(yoga);
 
+  const executor = buildHTTPExecutor({
+    endpoint: 'http://localhost:4000/graphql',
+    fetch: yoga.fetch,
+    File: yoga.fetchAPI.File,
+    FormData: yoga.fetchAPI.FormData,
+  });
+  asyncDisposableStack.use(executor);
   const client = createClient({
     url: 'http://localhost:4000/graphql',
-    exchanges: [
-      executorExchange(
-        buildHTTPExecutor({
-          endpoint: 'http://localhost:4000/graphql',
-          fetch: yoga.fetch,
-          File: yoga.fetchAPI.File,
-          FormData: yoga.fetchAPI.FormData,
-        }),
-      ),
-    ],
+    exchanges: [executorExchange(executor)],
   });
+
+  afterAll(() => asyncDisposableStack.disposeAsync());
 
   it('should handle queries correctly', async () => {
     const result = await client
@@ -118,7 +121,7 @@ describe('URQL Yoga Exchange', () => {
       expect(i).toBe(3);
     },
   );
-  testIf(!globalThis.Bun)('should handle file uploads correctly', async () => {
+  test('should handle file uploads correctly', async () => {
     const query = /* GraphQL */ `
       mutation readFile($file: File!) {
         readFile(file: $file)