+++ packages/loaders/url/src/defaultAsyncFetch.ts
@@ -1,6 +1,4 @@
 import { AsyncFetchFn } from '@graphql-tools/executor-http';
 import { fetch } from '@whatwg-node/fetch';
 
-export const defaultAsyncFetch: AsyncFetchFn = async (input, init) => {
-  return fetch(input, init);
-};
+export const defaultAsyncFetch: AsyncFetchFn = fetch;