+++ packages/utils/src/registerAbortSignalListener.ts
@@ -1,3 +1,4 @@
+import { fakeRejectPromise } from '@whatwg-node/promise-helpers';
 import { memoize1 } from './memoize.js';
 
 // AbortSignal handler cache to avoid the "possible EventEmitter memory leak detected"
@@ -32,8 +33,11 @@ export function registerAbortSignalListener(signal: AbortSignal, listener: () =>
 }
 
 export const getAbortPromise = memoize1(function getAbortPromise(signal: AbortSignal) {
+  // If the signal is already aborted, return a rejected promise
+  if (signal.aborted) {
+    return fakeRejectPromise(signal.reason);
+  }
   return new Promise<void>((_resolve, reject) => {
-    // If the signal is already aborted, return a rejected promise
     if (signal.aborted) {
       reject(signal.reason);
       return;