+++ prettier.config.mjs
@@ -3,4 +3,8 @@ import prettierConfig from '@theguild/prettier-config';
 export default {
   ...prettierConfig,
   plugins: [...prettierConfig.plugins, 'prettier-plugin-tailwindcss'],
+  importOrderParserPlugins: [
+    'explicitResourceManagement',
+    ...prettierConfig.importOrderParserPlugins,
+  ],
 };