+++ packages/mock/CHANGELOG.md
@@ -1,5 +1,101 @@
 # @graphql-tools/mock
 
+## 9.0.24
+
+### Patch Changes
+
+- [#7272](https://github.com/ardatan/graphql-tools/pull/7272)
+  [`803288c`](https://github.com/ardatan/graphql-tools/commit/803288c2d9c29fad539b3ee69bc532ac34f3f7c5)
+  Thanks [@XiNiHa](https://github.com/XiNiHa)! - Remove unneeded toString in MockStore
+
+- Updated dependencies
+  [[`22af985`](https://github.com/ardatan/graphql-tools/commit/22af98581e983079ec7e53677b905d3d03117524),
+  [`53db005`](https://github.com/ardatan/graphql-tools/commit/53db00540c2549748afdeeb65bb45a6c45ce57d3)]:
+  - @graphql-tools/utils@10.9.0
+  - @graphql-tools/schema@10.0.24
+
+## 9.0.23
+
+### Patch Changes
+
+- [#7210](https://github.com/ardatan/graphql-tools/pull/7210)
+  [`2443c23`](https://github.com/ardatan/graphql-tools/commit/2443c23c613163abe91995e69ae51f3fe86c00e6)
+  Thanks [@jdolle](https://github.com/jdolle)! - preserveResolvers uses resolved type name for
+  abstract types if available
+
+## 9.0.22
+
+### Patch Changes
+
+- [`6d69ede`](https://github.com/ardatan/graphql-tools/commit/6d69ede7a79521b5c132e33fcfc0f6ce2183ede7)
+  Thanks [@ardatan](https://github.com/ardatan)! - Fix prototype polluting assignment
+
+- Updated dependencies
+  [[`d123e26`](https://github.com/ardatan/graphql-tools/commit/d123e26b30b4febbbe1780bd32773b60e614dbf0)]:
+  - @graphql-tools/utils@10.8.6
+  - @graphql-tools/schema@10.0.23
+
+## 9.0.21
+
+### Patch Changes
+
+- [#7026](https://github.com/ardatan/graphql-tools/pull/7026)
+  [`ece7732`](https://github.com/ardatan/graphql-tools/commit/ece7732308b355972190f21372371ba3ae3a6296)
+  Thanks [@ardatan](https://github.com/ardatan)! - Prevent prototype polluting assignment
+
+- Updated dependencies
+  [[`90a717e`](https://github.com/ardatan/graphql-tools/commit/90a717e35a7e4e51da4fe747cb73544f24698fb7),
+  [`26518de`](https://github.com/ardatan/graphql-tools/commit/26518debfcb668e8feb5fb146271a13da92b778a)]:
+  - @graphql-tools/utils@10.8.5
+  - @graphql-tools/schema@10.0.22
+
+## 9.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  [[`155944b`](https://github.com/ardatan/graphql-tools/commit/155944b11e7ff1d8816ba3e9a4e4aa035c81f645)]:
+  - @graphql-tools/utils@10.8.4
+  - @graphql-tools/schema@10.0.21
+
+## 9.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  [[`4a2eb14`](https://github.com/ardatan/graphql-tools/commit/4a2eb14d0e3394d3dfbb3d83856500c5fb548285)]:
+  - @graphql-tools/utils@10.8.3
+  - @graphql-tools/schema@10.0.20
+
+## 9.0.18
+
+### Patch Changes
+
+- Updated dependencies
+  [[`3547bba`](https://github.com/ardatan/graphql-tools/commit/3547bbadd3ad4fa01a950fd685345f7c9c934e2f)]:
+  - @graphql-tools/utils@10.8.2
+  - @graphql-tools/schema@10.0.19
+
+## 9.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  [[`651a5dc`](https://github.com/ardatan/graphql-tools/commit/651a5dccb466b04f7fd16561cd264acd306e0711)]:
+  - @graphql-tools/utils@10.8.1
+  - @graphql-tools/schema@10.0.18
+
+## 9.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  [[`357b2ee`](https://github.com/ardatan/graphql-tools/commit/357b2ee6eef5e1f3b2b067728e9b167b1c2f82e5),
+  [`0a3e193`](https://github.com/ardatan/graphql-tools/commit/0a3e193e1ed70bd744375bfb4a10e515ecf59019),
+  [`357b2ee`](https://github.com/ardatan/graphql-tools/commit/357b2ee6eef5e1f3b2b067728e9b167b1c2f82e5)]:
+  - @graphql-tools/schema@10.0.17
+  - @graphql-tools/utils@10.8.0
+
 ## 9.0.15
 
 ### Patch Changes
@@ -704,7 +800,6 @@
 ### Major Changes
 
 - 24926654: Reworked to add state-full behavior to the library:
-
   - Breaking: mock functions does not receive resolver arguments anymore and can't return promise.
     Use `resolvers` option instead.
   - Breaking: when preserved, resolvers will not receive plain object returned by mock anymore as
@@ -720,7 +815,6 @@
 - be1a1575: ## Breaking Changes:
 
   #### Schema Generation and Decoration API (`@graphql-tools/schema`)
-
   - Resolver validation options should now be set to `error`, `warn` or `ignore` rather than `true`
     or `false`. In previous versions, some of the validators caused errors to be thrown, while some
     issued warnings. This changes brings consistency to validator behavior.
@@ -730,7 +824,6 @@
     `requireResolversToMatchSchema` is `error`, matching the previous behavior.
 
   #### Schema Delegation (`delegateToSchema` & `@graphql-tools/delegate`)
-
   - The `delegateToSchema` return value has matured and been formalized as an `ExternalObject`, in
     which all errors are integrated into the GraphQL response, preserving their initial path. Those
     advanced users accessing the result directly will note the change in error handling. This also
@@ -760,7 +853,6 @@
     parameters above.
 
   #### Remote Schemas & Wrapping (`wrapSchema`, `makeRemoteExecutableSchema`, and `@graphql-tools/wrap`)
-
   - `wrapSchema` and `generateProxyingResolvers` now only take a single options argument with named
     properties of type `SubschemaConfig`. The previously possible shorthand version with first
     argument consisting of a `GraphQLSchema` and second argument representing the transforms should
@@ -796,7 +888,6 @@
     `parseSelectionSet` function from `@graphql-tools/utils`.
 
   #### Schema Stitching (`stitchSchemas` & `@graphql-tools/stitch`)
-
   - `stitchSchemas`'s `mergeTypes` option is now true by default! This causes the `onTypeConflict`
     option to be ignored by default. To use `onTypeConflict` to select a specific type instead of
     simply merging, simply set `mergeTypes` to false.
@@ -817,13 +908,11 @@
     property name change for `onTypeConflict` from `schema` to `subschema`.
 
   #### Mocking (`addMocksToSchema` and `@graphql-tools/mock`)
-
   - Mocks returning objects with fields set as functions are now operating according to upstream
     graphql-js convention, i.e. these functions take three arguments, `args`, `context`, and `info`
     with `parent` available as `this` rather than as the first argument.
 
   #### Other Utilities (`@graphql-tools/utils`)
-
   - `filterSchema`'s `fieldFilter` will now filter _all_ fields across Object, Interface, and Input
     types. For the previous Object-only behavior, switch to the `objectFieldFilter` option.
   - Unused `fieldNodes` utility functions have been removed.
@@ -835,7 +924,6 @@
     removed from the `utils` package, as they are implemented elsewhere or no longer necessary.
 
   ## Related Issues
-
   - proxy all the errors: #1047, #1641
   - better error handling for merges #2016, #2062
   - fix typings #1614